pyproject.toml
[tool.poetry]
name = "scoring-api"
version = "0.1.0"
description = "Data Platform Scoring API"
authors = ["Artur Mullakhmetov <a.mullakhmetov@superdao.co>"]
packages = [
    { include = "app" },
]
[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.93.0"
pydantic = {extras = ["dotenv"], version = "^1.10.6"}
uvicorn = "0.20.0"
flake8-pyproject = "^1.2.2"
httpx = "^0.23.3"
asyncpg = "^0.27.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.5.post1"}
pytest-asyncio = "^0.20.3"
mock = "^5.0.1"
sentry-sdk = {extras = ["fastapi"], version = "^1.18.0"}
prometheus-client = "^0.16.0"
gunicorn = "^20.1.0"
python-multipart = "^0.0.6"
google-cloud-storage = "^2.8.0"
types-google-cloud-ndb = "^2.1.0.6"
types-python-dateutil = "^2.8.19.12"
python-dateutil = "^2.8.2"
eth-hash = {extras = ["pycryptodome"], version = "^0.5.1"}
eth-utils = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.2.2"
flake8 = "6.0.0"
black = "23.1.0"
flake8-black = "^0.3.6"
mypy = "^1.1.1"
isort = "^5.12.0"
bandit = {extras = ["toml"], version = "^1.7.4"}
ipdb = "^0.13.11"
aiosqlite = "^0.18.0"
psycopg2-binary = "^2.9.5"
locust = "^2.15.1"
pglast = "^5.2"
testing-postgresql = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
extend-ignore = [
    # See https://github.com/PyCQA/pycodestyle/issues/373
    "E203",
]

[tool.black]
skip-string-normalization = true


[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = false
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_equality = true
no_implicit_optional = true
show_error_codes = true
plugins = ["sqlalchemy.ext.mypy.plugin"]
exclude = [
    'tests', 
    '^app\/generate_db_scripts\.py$',
]

[tool.isort]
profile = "black"

[tool.bandit]
exclude_dirs = ["tests"]

[tool.pytest.ini_options]
markers = [
    "db_integration: marks tests as requiring database (deselect with '-m \"not db_integration\"')",
]